pyproject.toml
[project]
name = "openai-basics"
version = "0.1.0"
description = ""
authors = [
    {name = "priyank1725",email = "priyank000galani@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "openai (>=1.100.2,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "textblob (>=0.19.0,<0.20.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-google-genai (>=2.1.9,<3.0.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "chromadb (>=1.0.20,<2.0.0)",
    "tiktoken (>=0.11.0,<0.12.0)",
    "pillow (>=11.3.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
